Initial Commit
This is still very much a work in progress but should hold up well.
If you're only interested in the OCR functionality I recomend you
take a look in application/classes for the actual api and database
classes that actually handle all the database logic. I may release
that as a seperate repository at some stage, but for now this works
fine.